refactor(gui-kitchen-sink): DDDに基づく構造改善
## App.tsxを分割してドメイン駆動の構造に
- 380行のApp.tsxをセクション単位で分割
- ドメイン概念を反映したコンポーネント名
  ProjectActions, ProjectConfiguration, ViewPreferences,
  TeamRoster, DocumentUpload

## 関心別ディレクトリ構成
- components/layout/: 汎用レイアウト（Section）
- components/theme/: テーマ・ビジュアル要素（PageWithTheme, icons）
- components/dialogs/: ダイアログ（SettingsDialog）
- components/直下: ドメイン機能コンポーネント
- sample-data/: デモ用サンプルデータ

## ドメイン層とUI層の分離
- src/models/: 型定義とバリデーションのみ（Framework, Plan, Layout, Member）
- UI表示用データはコンポーネント内に定義
