+++ packages/gui-kitchen-sink/gui/sample-data/members.ts
@@ -1,0 +1,8 @@
+import type { Member } from '../../src/models/Member';
+
+export const SAMPLE_MEMBERS: Member[] = [
+  { id: 1, name: 'Alice Johnson', role: 'Product Manager', experience: 8 },
+  { id: 2, name: 'Mia Chen', role: 'Product Designer', experience: 6 },
+  { id: 3, name: 'Liam Patel', role: 'Frontend Engineer', experience: 5 },
+  { id: 4, name: 'Chloe Smith', role: 'Backend Engineer', experience: 7 },
+];
