+++ packages/gui-kitchen-sink/src/models/Layout.ts
@@ -1,0 +1,5 @@
+export type Layout = 'dashboard' | 'timeline' | 'board';
+
+export const Layout = {
+  guard: (x: unknown): x is Layout => ['dashboard', 'timeline', 'board'].includes(x as string),
+};
