+++ packages/gui-kitchen-sink/src/models/Member.ts
@@ -1,0 +1,6 @@
+export type Member = {
+  id: number;
+  name: string;
+  role: string;
+  experience: number;
+};
