+++ packages/gui-kitchen-sink/src/models/Project.ts
@@ -1,12 +1,7 @@
-export type Framework = 'react' | 'svelte' | 'vue';
-export const Framework = {
-  guard: (x: unknown): x is Framework => ['react', 'svelte', 'vue'].includes(x as string),
-};
+import type { Framework } from './Framework';
+import type { Plan } from './Plan';
 
-export type Plan = 'starter' | 'team' | 'enterprise';
-export const Plan = {
-  guard: (x: unknown): x is Plan => ['starter', 'team', 'enterprise'].includes(x as string),
-};
+export type { Framework, Plan };
 
 export type Project = {
   projectName: string;
